+++ app/banking/templates/banking/home.html
@@ -48,6 +48,34 @@
         <h1>{{ bank_settings.bank_name }}</h1>
         <p>Monitor balances and account activity across every banking channel.</p>
       </header>
+
+      {% if missing_accounts %}
+        {% set missing_titles = missing_accounts | map('capitalize') | list %}
+        {% set missing_phrase = missing_titles | join(' and ') %}
+        <section class="bank-onboarding" aria-label="Bank account onboarding">
+          <div class="bank-onboarding__copy">
+            {% if not has_open_accounts %}
+              <h2>Welcome to {{ bank_settings.bank_name }}</h2>
+              <p>
+                Open your first checking or savings account with {{ available_cash }} ready in cash to allocate.
+              </p>
+            {% else %}
+              <h2>{{ bank_settings.bank_name }} account options</h2>
+              <p>
+                Your {{ missing_phrase }} account is still closed. Move cash into the bank to unlock more tools.
+              </p>
+            {% endif %}
+          </div>
+          <div class="bank-onboarding__actions">
+            <button type="button" class="cta" data-open-account>Open Account</button>
+          </div>
+          <p class="bank-onboarding__footnote">
+            Minimum deposits — Checking: {{ account_opening_requirements.checking }}, Savings:
+            {{ account_opening_requirements.savings }}
+          </p>
+        </section>
+      {% endif %}
+
       <section class="account-activity" aria-label="Account activity">
         <header class="account-activity__header">
           <h2>Account Activity</h2>
@@ -92,7 +120,7 @@
             <p class="account-activity__description">Track the latest totals for each account.</p>
             {% if not has_open_accounts %}
               <p class="account-activity__empty">
-                No banking accounts are open yet. Use the Open Account action on the home page to get started.
+                No banking accounts are open yet. Use the Open Account action above to get started.
               </p>
             {% endif %}
             <div class="account-grid">
@@ -129,5 +157,80 @@
         </div>
       </section>
     </section>
+
+    <div class="modal" id="account-opening-modal" role="dialog" aria-modal="true" aria-labelledby="account-opening-title" hidden>
+      <div class="modal__dialog">
+        <header class="modal__header">
+          <h2 id="account-opening-title">Open a banking account</h2>
+          <button type="button" class="modal__close" data-close-modal aria-label="Close">×</button>
+        </header>
+        <form id="account-opening-form" class="modal__form">
+          <p class="modal__intro">
+            Choose which accounts to open and decide how much to deposit from cash. You currently have {{ available_cash }} on hand.
+          </p>
+          <fieldset class="modal__fieldset">
+            <legend>Select account types</legend>
+            <label class="modal__option">
+              <input type="checkbox" value="checking" data-account-option {% if 'checking' not in missing_accounts %}disabled{% endif %} />
+              <span class="modal__option-label">
+                Checking Account
+                <span class="modal__option-note">
+                  {% if 'checking' not in missing_accounts %}
+                    Already open
+                  {% else %}
+                    Minimum deposit {{ account_opening_requirements.checking }}
+                  {% endif %}
+                </span>
+              </span>
+            </label>
+            <div class="modal__deposit" data-deposit-section="checking" hidden>
+              <label>
+                <span>Deposit amount</span>
+                <div class="modal__input">
+                  <span class="modal__input-prefix">$</span>
+                  <input type="number" step="0.01" min="0" data-deposit-input="checking" />
+                </div>
+              </label>
+            </div>
+            <label class="modal__option">
+              <input type="checkbox" value="savings" data-account-option {% if 'savings' not in missing_accounts %}disabled{% endif %} />
+              <span class="modal__option-label">
+                Savings Account
+                <span class="modal__option-note">
+                  {% if 'savings' not in missing_accounts %}
+                    Already open
+                  {% else %}
+                    Minimum deposit {{ account_opening_requirements.savings }}
+                  {% endif %}
+                </span>
+              </span>
+            </label>
+            <div class="modal__deposit" data-deposit-section="savings" hidden>
+              <label>
+                <span>Deposit amount</span>
+                <div class="modal__input">
+                  <span class="modal__input-prefix">$</span>
+                  <input type="number" step="0.01" min="0" data-deposit-input="savings" />
+                </div>
+              </label>
+            </div>
+          </fieldset>
+          <div class="modal__feedback" data-modal-feedback hidden></div>
+          <div class="modal__actions">
+            <button type="button" class="cta secondary" data-close-modal>Cancel</button>
+            <button type="submit" class="cta" data-submit-button>Open Selected Accounts</button>
+          </div>
+        </form>
+      </div>
+    </div>
+    <div class="modal-backdrop" data-modal-backdrop hidden></div>
+
+    <script type="application/json" id="account-opening-config">
+      {{ {
+        "available_cash": available_cash_value,
+        "requirements": account_opening_requirements_value,
+        "endpoint": url_for('banking.api_open_accounts')
+      } | tojson }}
+    </script>
   </div>
 {% endblock %}
