+++ app/index/static/styles/home.css
@@ -124,191 +124,3 @@
   border-left-color: var(--accent);
   color: var(--accent);
 }
-
-.bank-onboarding {
-  margin: clamp(1.5rem, 4vw, 2.5rem) 0;
-  padding: clamp(1.5rem, 4vw, 2.25rem);
-  border-radius: 1.5rem;
-  border: 1px solid rgba(148, 163, 184, 0.25);
-  background: linear-gradient(135deg, rgba(59, 130, 246, 0.1), rgba(15, 23, 42, 0.85));
-  display: grid;
-  gap: 1.25rem;
-  align-items: center;
-  color: #f8fafc;
-}
-
-.bank-onboarding__copy h2 {
-  margin: 0 0 0.35rem;
-  font-size: clamp(1.75rem, 3vw, 2.2rem);
-}
-
-.bank-onboarding__copy p {
-  margin: 0;
-  line-height: 1.6;
-}
-
-.bank-onboarding__actions {
-  display: flex;
-  align-items: center;
-}
-
-.bank-onboarding__footnote {
-  margin: 0;
-  font-size: 0.85rem;
-  color: rgba(226, 232, 240, 0.8);
-}
-
-.modal[hidden],
-.modal-backdrop[hidden] {
-  display: none;
-}
-
-.modal {
-  position: fixed;
-  inset: 0;
-  display: grid;
-  place-items: center;
-  z-index: 1000;
-}
-
-.modal__dialog {
-  width: min(90%, 520px);
-  background: #ffffff;
-  border-radius: 1.25rem;
-  border: 1px solid rgba(148, 163, 184, 0.35);
-  padding: clamp(1.5rem, 4vw, 2.25rem);
-  display: grid;
-  gap: 1.5rem;
-  color: var(--text);
-}
-
-.modal__header {
-  display: flex;
-  justify-content: space-between;
-  align-items: start;
-  gap: 1rem;
-}
-
-.modal__header h2 {
-  margin: 0;
-  font-size: 1.6rem;
-}
-
-.modal__close {
-  background: transparent;
-  border: none;
-  color: var(--muted);
-  font-size: 1.5rem;
-  cursor: pointer;
-}
-
-.modal__form {
-  display: grid;
-  gap: 1.5rem;
-}
-
-.modal__intro {
-  margin: 0;
-  color: var(--muted);
-  line-height: 1.6;
-}
-
-.modal__fieldset {
-  border: 1px solid #dfe4ec;
-  border-radius: 1rem;
-  padding: 1.25rem;
-  display: grid;
-  gap: 1rem;
-}
-
-.modal__fieldset legend {
-  padding: 0 0.5rem;
-  font-weight: 700;
-  color: var(--text);
-}
-
-.modal__option {
-  display: flex;
-  align-items: flex-start;
-  gap: 0.75rem;
-  font-weight: 600;
-}
-
-.modal__option input {
-  margin-top: 0.25rem;
-}
-
-.modal__option-label {
-  display: grid;
-  gap: 0.25rem;
-}
-
-.modal__option-note {
-  font-size: 0.85rem;
-  font-weight: 400;
-  color: var(--muted);
-}
-
-.modal__deposit {
-  display: grid;
-  gap: 0.5rem;
-}
-
-.modal__deposit label {
-  display: grid;
-  gap: 0.35rem;
-  font-weight: 600;
-}
-
-.modal__input {
-  display: flex;
-  align-items: center;
-  border: 1px solid #dfe4ec;
-  border-radius: 0.75rem;
-  padding: 0.5rem 0.75rem;
-  gap: 0.5rem;
-  background: #f8fafc;
-}
-
-.modal__input-prefix {
-  font-weight: 600;
-  color: var(--muted);
-}
-
-.modal__input input {
-  border: none;
-  background: transparent;
-  font-size: 1rem;
-  width: 100%;
-}
-
-.modal__feedback {
-  padding: 0.75rem 1rem;
-  border-radius: 0.75rem;
-  font-weight: 600;
-}
-
-.modal__feedback[data-state="error"] {
-  background: rgba(185, 28, 28, 0.1);
-  color: var(--error);
-  border: 1px solid rgba(185, 28, 28, 0.2);
-}
-
-.modal__feedback[data-state="success"] {
-  background: rgba(21, 128, 61, 0.1);
-  color: var(--success);
-  border: 1px solid rgba(21, 128, 61, 0.2);
-}
-
-.modal__actions {
-  display: flex;
-  justify-content: flex-end;
-  gap: 1rem;
-}
-
-.modal-backdrop {
-  position: fixed;
-  inset: 0;
-  background: rgba(15, 23, 42, 0.6);
-  z-index: 999;
-}
